deeplearning_ai/acp/my_acp_project/pyproject.toml
[project]
name = "my-acp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "acp-sdk>=0.9.1",
    "colorama>=0.4.6",
    "crewai>=0.121.0",
    "crewai-tools>=0.45.0",
    "duckduckgo-search>=8.0.2",
    "load-dotenv>=0.1.0",
    "markdownify>=1.1.0",
    "mcp>=1.9.1",
    "nest-asyncio>=1.6.0",
    "ollama>=0.5.1",
    "requests>=2.32.3",
    "smolagents[mcp]>=1.16.1",
]
